Embeddings/requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.2
pandas==1.5.3
numpy==1.24.2
